updat:task 5 list typing

+++ 0x00-python_variable_annotations/5-sum_list.py
@@ -3,7 +3,7 @@
 from typing import List
 
 
-def sum_list(input_list: List[int]) -> float:
+def sum_list(input_list: List[float]) -> float:
     """sum_list takes a list input_list of
     floats as argument and returns their
     sum as a float."""
